requirements.txt
requests
validators
dnspython
tqdm
